跨平台兼容性修复：支持 Linux/macOS 构建
- MainWindow.cpp: 将 Windows API 调用包裹在 #ifdef Q_OS_WIN 中
  (AllocConsole, FreeConsole, WTSRegister, SetWindowPos, FindWindowA)
- MainWindow.h: 添加缺失的 QShortcut 头文件
- BroadcastReceiver: 添加 POSIX socket 替代 WinSock 实现
- SystemMonitor: 添加 Linux /proc 文件系统实现替代 Windows API
- hardware_decoder: 添加 VAAPI/VDPAU (Linux), VideoToolbox (macOS) 支持
- packet_queue.h: 替换不可用的 QtLogging 为 QLoggingCategory
- ApplicationSettings: 修复 int64_t 到 QJsonValue 的歧义转换
- DrawWidget: 替换 MSVC __super 为标准 QWidget::
- VideoThread: 修复 SPDLOG_LOGGER_DEBUG 缺失 logger 参数, localtime_s->localtime_r
- VideoWidgetBase: 改用 QOpenGLWidget 基类以支持 GL 方法 override
- OpenGLVideoWidget: 移除多余的 QOpenGLWidget 构造

+++ WZMediaPlay/ApplicationSettings.cpp
@@ -543,7 +543,7 @@ void ApplicationSettings::write_PlayList()
     QJsonObject rootObj;
     rootObj.insert("playlist_current_index", GlobalDef::getInstance()->PLAY_LIST_DATA.playlist_current_index);
     rootObj.insert("playlist_current_video", GlobalDef::getInstance()->PLAY_LIST_DATA.playlist_current_video);
-    rootObj.insert("playlist_video_play_time", GlobalDef::getInstance()->PLAY_LIST_DATA.playlist_video_play_time);
+    rootObj.insert("playlist_video_play_time", static_cast<qint64>(GlobalDef::getInstance()->PLAY_LIST_DATA.playlist_video_play_time));
     QJsonArray playListJsonArray;
     for (int i = 0; i < GlobalDef::getInstance()->PLAY_LIST_DATA.play_list.size(); ++i) {
         QJsonObject itemObj;
